Installed react router. Added home company and search web pages. Refactored all state and logic to the search page. Added links to appropriate web pages
