+++ front-end/front-end/src/Components/Hero/Hero.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import hero from './hero.png'
+import { Link } from 'react-router-dom'
 
 interface Props {
 
@@ -18,12 +19,12 @@ const Hero = (props: Props) => {
             news.
           </p>
           <div className="mx-auto lg:mx-0">
-            <a
-              href=""
+            <Link
+              to="/search"
               className="py-5 px-10 text-2xl font-bold text-white bg-lightGreen rounded lg:py-4 hover:opacity-70"
             >
               Get Started
-            </a>
+            </Link>
           </div>
         </div>
         <div className="mb-24 mx-auto md:w-180 md:px-10 lg:mb-0 lg:w-1/2">
