+++ front-end/front-end/src/Components/Navbar/Navbar.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import logo from './logo.png';
+import to from 'react-router-dom';
+import { Link } from 'react-router-dom';
 
 interface Props {}
 
@@ -8,11 +10,13 @@ const Navbar = (props: Props) => {
     <nav className="relative container mx-auto p-6">
       <div className="flex items-center justify-between">
         <div className="flex items-center space-x-20">
-          <img src={logo} alt="" />
+            <Link to="/"> 
+                <img src={logo} alt="" />
+            </Link>
           <div className="hidden font-bold lg:flex">
-            <a href="" className="text-black hover:text-darkBlue">
-              Dashboard
-            </a>
+            <Link to="/search" className="text-black hover:text-darkBlue">
+              Search
+            </Link>
           </div>
         </div>
         <div className="hidden lg:flex items-center space-x-6 text-back">
