+++ front-end/front-end/src/Pages/CompanyPage/CompanyPage.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+interface Props {
+    
+}
+
+const CompanyPage = (props: Props) => {
+  return (
+    <div>CompanyPage</div>
+  )
+}
+
+export default CompanyPage
